+++ test/Apache.Arrow.Operations.Tests/Json/VariantJsonTests.cs
