+++ test/Apache.Arrow.Variant.Tests/VariantSqlDecimalTests.cs
@@ -16,8 +16,6 @@
 using System;
 using System.Collections.Generic;
 using System.Data.SqlTypes;
-using System.Text.Json;
-using Apache.Arrow.Operations.Json;
 using Xunit;
 
 namespace Apache.Arrow.Variant.Tests
@@ -279,39 +277,6 @@ namespace Apache.Arrow.Variant.Tests
         }
 
         // ---------------------------------------------------------------
-        // JSON writer: large Decimal16 produces valid JSON number
-        // ---------------------------------------------------------------
-
-        [Fact]
-        public void JsonWriter_LargeDecimal16_WritesValidJson()
-        {
-            byte[] value = new byte[]
-            {
-                Decimal16Header,
-                0x00, // scale = 0
-                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                0x00, 0x00, 0x00, 0x00, 0x01, 0x00, 0x00, 0x00
-            };
-
-            string json = VariantJsonWriter.ToJson(MinimalMetadata, value);
-            Assert.Equal("79228162514264337593543950336", json);
-        }
-
-        // ---------------------------------------------------------------
-        // JSON converter: VariantValue with SqlDecimal storage writes valid JSON
-        // ---------------------------------------------------------------
-
-        [Fact]
-        public void JsonConverter_SqlDecimalStorage_WritesValidJson()
-        {
-            SqlDecimal sd = SqlDecimal.Parse("99999999999999999999999999999999999999");
-            VariantValue vv = VariantValue.FromSqlDecimal(sd);
-
-            string json = VariantJsonWriter.ToJson(vv);
-            Assert.Equal("99999999999999999999999999999999999999", json);
-        }
-
-        // ---------------------------------------------------------------
         // Round-trip with materialization
         // ---------------------------------------------------------------
 
@@ -501,37 +466,6 @@ namespace Apache.Arrow.Variant.Tests
         }
 
         // ---------------------------------------------------------------
-        // JSON writer for negative large Decimal16
-        // ---------------------------------------------------------------
-
-        [Fact]
-        public void JsonWriter_NegativeLargeDecimal16_WritesValidJson()
-        {
-            SqlDecimal sd = SqlDecimal.Parse("-79228162514264337593543950336");
-            VariantValue vv = VariantValue.FromSqlDecimal(sd);
-            VariantBuilder builder = new VariantBuilder();
-            (byte[] metadata, byte[] value) = builder.Encode(vv);
-
-            string json = VariantJsonWriter.ToJson(metadata, value);
-            Assert.Equal("-79228162514264337593543950336", json);
-        }
-
-        // ---------------------------------------------------------------
-        // JSON writer for small Decimal16 (decimal path)
-        // ---------------------------------------------------------------
-
-        [Fact]
-        public void JsonWriter_SmallDecimal16_WritesNumberValue()
-        {
-            VariantValue vv = VariantValue.FromDecimal16(42.5m);
-            VariantBuilder builder = new VariantBuilder();
-            (byte[] metadata, byte[] value) = builder.Encode(vv);
-
-            string json = VariantJsonWriter.ToJson(metadata, value);
-            Assert.Equal("42.5", json);
-        }
-
-        // ---------------------------------------------------------------
         // GetSqlDecimal on non-decimal primitive throws
         // ---------------------------------------------------------------
 
@@ -604,25 +538,6 @@ namespace Apache.Arrow.Variant.Tests
             Assert.False(elements[1].IsSqlDecimalStorage);
             Assert.Equal(42.5m, elements[1].AsDecimal());
             Assert.Equal(1.23m, elements[2].AsDecimal());
-        }
-
-        // ---------------------------------------------------------------
-        // JSON converter serialization of nested SqlDecimal
-        // ---------------------------------------------------------------
-
-        [Fact]
-        public void JsonConverter_ObjectWithSqlDecimal_WritesValidJson()
-        {
-            SqlDecimal sd = SqlDecimal.Parse("99999999999999999999999999999999999999");
-            VariantValue obj = VariantValue.FromObject(new Dictionary<string, VariantValue>
-            {
-                { "val", VariantValue.FromSqlDecimal(sd) },
-            });
-
-            JsonSerializerOptions options = new JsonSerializerOptions();
-            options.Converters.Add(VariantJsonConverter.Instance);
-            string json = JsonSerializer.Serialize(obj, options);
-            Assert.Equal("{\"val\":99999999999999999999999999999999999999}", json);
         }
 
         // ---------------------------------------------------------------
